+++ unsloth_compiled_cache/GroupNorm.py
@@ -58,7 +58,7 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 from typing import Any, List, Optional, Tuple, Union, Dict, Set, Callable
-from transformers.models.siglip.modeling_siglip import (np)
+from transformers.models.gemma3n.modeling_gemma3n import (F)
 
 def forward(self, input: Tensor) -> Tensor:
-    return F.group_norm(input, self.num_groups, self.weight, self.bias, self.eps).to(input.dtype)
+    return F.group_norm(input, self.num_groups, self.weight, self.bias, self.eps).to(input.dtype).to(input.dtype)
